refactor(integrations): hoist delivery-readiness helpers into platform/scheduler (#4164)
