+++ tests/integrations/sentry/test_digest_delivery.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 import pytest
 
-from integrations.sentry.digest_delivery import (
+from platform.scheduler.delivery_readiness import (
     any_digest_delivery_ready,
     delivery_provider_ready,
     digest_delivery_setup_hint,
@@ -17,11 +17,11 @@ from platform.scheduler.types import Provider
 class TestDigestDeliveryReadiness:
     def test_telegram_ready_when_token_present(self, monkeypatch: pytest.MonkeyPatch) -> None:
         monkeypatch.setattr(
-            "integrations.sentry.digest_delivery.resolve_telegram_credentials",
+            "platform.scheduler.delivery_readiness.resolve_telegram_credentials",
             lambda _params: {"bot_token": "token"},
         )
         monkeypatch.setattr(
-            "integrations.sentry.digest_delivery.resolve_slack_credentials",
+            "platform.scheduler.delivery_readiness.resolve_slack_credentials",
             lambda _params: {},
         )
         assert telegram_delivery_ready() is True
@@ -30,11 +30,11 @@ class TestDigestDeliveryReadiness:
 
     def test_slack_ready_with_webhook(self, monkeypatch: pytest.MonkeyPatch) -> None:
         monkeypatch.setattr(
-            "integrations.sentry.digest_delivery.resolve_telegram_credentials",
+            "platform.scheduler.delivery_readiness.resolve_telegram_credentials",
             lambda _params: {},
         )
         monkeypatch.setattr(
-            "integrations.sentry.digest_delivery.resolve_slack_credentials",
+            "platform.scheduler.delivery_readiness.resolve_slack_credentials",
             lambda _params: {"webhook_url": "https://hooks.slack.com/services/x"},
         )
         assert slack_delivery_ready() is True
@@ -42,11 +42,11 @@ class TestDigestDeliveryReadiness:
 
     def test_none_ready(self, monkeypatch: pytest.MonkeyPatch) -> None:
         monkeypatch.setattr(
-            "integrations.sentry.digest_delivery.resolve_telegram_credentials",
+            "platform.scheduler.delivery_readiness.resolve_telegram_credentials",
             lambda _params: {},
         )
         monkeypatch.setattr(
-            "integrations.sentry.digest_delivery.resolve_slack_credentials",
+            "platform.scheduler.delivery_readiness.resolve_slack_credentials",
             lambda _params: {},
         )
         assert any_digest_delivery_ready() is False
